specify external top-level functions
