+++ prusti-interface/src/specs/external.rs
@@ -59,17 +59,22 @@ impl<'tcx> intravisit::Visitor<'tcx> for ExternSpecResolver<'tcx> {
 
     fn visit_expr(&mut self, ex: &'tcx rustc_hir::Expr<'tcx>) {
         if let rustc_hir::ExprKind::Call(ref callee_expr, ref arguments) = ex.kind {
-            if self.current_def_id.is_some() {
-                let def_id = self.tcx.typeck(callee_expr.hir_id.owner).type_dependent_def_id(callee_expr.hir_id);
-                assert!(def_id.is_some());
-                if self.extern_fn_map.contains_key(&def_id.unwrap()) {
-                    self.tcx.ty_error_with_message(ex.span.source_callsite(),
-                                                   &format!("Duplicate specification for {:?}",
-                                                            def_id.unwrap()));
-                    self.current_def_id = None;
-                } else {
-                    self.extern_fn_map.insert(def_id.unwrap(), self.current_def_id.take().unwrap());
-                }
+            if self.current_def_id.is_none() {
+                intravisit::walk_expr(self, ex);
+                return;
+            }
+            if let rustc_hir::ExprKind::Path(ref qself) = callee_expr.kind {
+                let res = self.tcx.typeck(callee_expr.hir_id.owner).qpath_res(qself, callee_expr.hir_id);
+                if let rustc_hir::def::Res::Def(_, def_id) = res {
+                    if self.extern_fn_map.contains_key(&def_id) {
+                        self.tcx.ty_error_with_message(ex.span.source_callsite(),
+                                                       &format!("Duplicate specification for {:?}",
+                                                                def_id));
+                        self.current_def_id = None;
+                    } else {
+                        self.extern_fn_map.insert(def_id, self.current_def_id.take().unwrap());
+                    }
+                };
             }
         }
         intravisit::walk_expr(self, ex);
